refactor: move redis stream contract
